+++ RELEASES.md
@@ -23,6 +23,7 @@ Upcoming
 - Upgrade CmisSender to support 'browser' bindings, support proxys and secure connections
 - Refactor HttpSender and WebServiceSender and (LSM) upgrade HttpClient from version 3 to version 4. Better (Mtom-)Multipart implementation and support for other http implementations using the HttpSenderBase and HttpResponseHandler
 - Fix NPE DumpIbisConsole
+- Refactor CheckReload job